rename APIExceptions file to exceptions
